mod some file

+++ m0/Libft/libft.h
@@ -17,6 +17,7 @@
 # include <stdlib.h>
 # include <string.h>
 # include <unistd.h>
+# include <ctype.h>
 
 // Characters
 int	ft_isalnum(int c);
@@ -30,7 +31,7 @@ int	ft_toupper(int c);
 // Memory
 void	ft_bzero(void *s, size_t n);
 void	*ft_memchr(const void *str, int c, size_t n);
-int	    ft_memcmp(const void *s1, const void *s2, int n);
+int	    ft_memcmp(const void *s1, const void *s2, size_t n);
 void	*ft_memcpy(void *dest, const void *src, size_t n);
 void	*ft_memmove(void *dest, const void *src, size_t n);
 void	*ft_memset(void *str, int c, size_t n);
